Restructure front-end config to use defaults for unspecified fields
...for any fields not specified by the project's configuration.
